Modularize FFmpeg conversion flow
