+++ app/build.gradle.kts
@@ -41,6 +41,9 @@ android {
 
 dependencies {
 
+    implementation(project(":feature:home"))
+    implementation(project(":core:ffmpeg"))
+
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.lifecycle.runtime.ktx)
     implementation(libs.androidx.activity.compose)
@@ -49,9 +52,6 @@ dependencies {
     implementation(libs.androidx.compose.ui.graphics)
     implementation(libs.androidx.compose.ui.tooling.preview)
     implementation(libs.androidx.compose.material3)
-    implementation("io.github.jamaismagic.ffmpeg:ffmpeg-kit-lts-16kb:6.0.1") // Switched to jamaismagic fork
-    implementation("io.coil-kt:coil-gif:2.6.0")
-    implementation("io.coil-kt:coil-compose:2.6.0")
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
     androidTestImplementation(libs.androidx.espresso.core)
