+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,8 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.Giffer2">
+        android:theme="@style/Theme.Giffer2"
+        android:name=".GifVisionApplication">
         <activity
             android:name=".MainActivity"
             android:exported="true"
